requirement.txt
asgiref==3.11.0
dj-database-url==3.0.1
Django==5.2.8
djangorestframework==3.16.1
drf-yasg==1.21.8
django-filter==25.1
djoser==2.3.1
gunicorn==23.0.0
packaging==25.0
psycopg2-binary==2.9.11
python-decouple==3.8
sqlparse==0.5.4
tzdata==2025.2
whitenoise==6.9.0